requirements.txt

batchgenerators==0.23
brotlipy==0.7.0
bunch==1.0.1
certifi==2021.10.8


click==8.1.3
connected-components-3d==3.10.0

docker-pycreds==0.4.0
einops==0.4.1
fastremap==1.13.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27

imageio==2.17.0
joblib==1.1.0
linecache2==1.0.0
lmdb==1.3.0
loguru==0.6.0
mkl-fft==1.3.1

mkl-service==2.4.0
networkx==2.8
nibabel==3.2.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pathtools==0.1.2
Pillow==9.1.0
prefetch-generator==1.0.1
promise==2.3
protobuf==3.20.1
psutil==5.9.0

pyparsing==3.0.8

python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0

ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
SimpleITK==2.0.2

smmap==5.0.0
threadpoolctl==3.1.0
tifffile==2022.4.22
timm==0.5.4

tqdm==4.64.0
traceback2==1.4.0

unittest2==1.1.0

wandb==0.12.15
yacs==0.1.8
